basic layout for patient dashboard
