+++ controllers/index.js
@@ -4,14 +4,14 @@ const router = require('express').Router();
 // require routes
 const apiRoutes = require('./api');
 const homeRoutes = require('./home-routes');
-const dashboardPatRoutes = require('./dashboard-pat-routes');
+const dashboardPatRoutes = require('./dashpatient-routes');
 
 
 // assign routes
 router.use('/api', apiRoutes);
 router.use('/', homeRoutes);
 router.use('/signup', homeRoutes);
-router.use('/dashboard-pat', dashboardPatRoutes);
+router.use('/dashpatient', dashboardPatRoutes);
 
 // export all router routes
 module.exports = router;
