+++ public/css/stylesheet.css
@@ -29,12 +29,20 @@ body {
     box-shadow: 0 0 10px rgba(100,60,60,0.5);
     border-radius: 20px;
     background: var(--secondary-color);
-    background: rgba(235, 228, 228, 0.25);
+    background: rgba(235, 228, 228, 0.75);
 }
 
 p {
     color: var(--primary-color);
     font-weight: bolder;
+}
+
+.demo {
+    background-image: url(/images/calendar.jpg);
+    background-size: cover;
+    padding-left: 90%;
+    width: 500px;
+    height: 800px;
 }
 
 a {
