+++ views/dashpatient.hbs
@@ -1,6 +1,6 @@
 <link rel="stylesheet" href="/css/secondary.css">
 
-<header class="row p-3 border-bottom border-5">
+<header class="row p-3 border-bottom border-info border-5">
   <h1 class="col-10">
     <a href="/">
       Med Planner
@@ -44,7 +44,7 @@
 </section>
 
    
-<section>
+<section class="demo">
   {{!-- appointment calender will go here --}}
 </section>
